requirements.txt
praw==7.7.1
transformers==4.52.4
torch==2.7.1
datasets==3.6.0
python-dotenv==1.0.0
requests==2.32.3
Flask==3.1.1